+++ packages/api/test/nfts-store.spec.js
@@ -80,12 +80,6 @@ describe('NFT store', () => {
         status: 'PinQueued',
         service: 'IpfsCluster3',
       },
-      {
-        content_cid:
-          'bafyreicnwbboevx6g6fykitf4nebz2kqgkqz35qvlnlcgfulhrris66m6i',
-        status: 'PinQueued',
-        service: 'Pinata',
-      },
     ])
   })
 
@@ -144,12 +138,6 @@ describe('NFT store', () => {
           'bafyreibubvuqeh5lm2ccose7uq54wweeajtjuhcj4pciogeq2lhbw2a23y',
         status: 'PinQueued',
         service: 'IpfsCluster3',
-      },
-      {
-        content_cid:
-          'bafyreibubvuqeh5lm2ccose7uq54wweeajtjuhcj4pciogeq2lhbw2a23y',
-        status: 'PinQueued',
-        service: 'Pinata',
       },
     ])
   })